+++ platforms/fabric-1.19/src/main/java/org/anti_ad/mc/common/vanilla/Vanilla.kt
@@ -79,10 +79,10 @@ object Vanilla {
         mc().interactionManager ?: error("mc.interactionManager is not initialized! Probably not in game")
 
     fun recipeBook() = player().recipeBook ?: throw AssertionError("unreachable")
-
+/*
     @Suppress("FunctionName")
     fun ClientPlayerEntity.`(sendChatMessage)`(msg: String)  = sendChatMessage(msg)
-
+*/
     val px
         get() = playerNullable()?.x.orDefault { 0.0 }
     val py