+++ platforms/forge-1.15/src/main/java/org/anti_ad/mc/ipnext/item/ItemTypeExtensions.kt
@@ -136,6 +136,18 @@ inline val ItemType.durability: Int
 
 inline val ItemType.isBucket: Boolean
     get() = item is BucketItem || item is MilkBucketItem || item is FishBucketItem
+
+inline val ItemType.isFullBucket: Boolean
+    get() = (item is BucketItem && item != Items.BUCKET) || item is MilkBucketItem || item is FishBucketItem
+
+inline val ItemType.isEmptyBucket: Boolean
+    get() {
+        return item == Items.BUCKET
+    }
+
+inline val ItemType.isHoneyBottle: Boolean
+    get() = item == Items.HONEY_BOTTLE
+
 inline val ItemType.isStew: Boolean
     get() = item is MushroomStewItem || item is SuspiciousStewItem // SoupItem = MushroomStewItem
 //endregion