+++ shared-sources/src/main/java/org/anti_ad/mc/ipnext/item/ItemStack.kt
@@ -30,6 +30,10 @@ sealed class ItemStack {
     inline val overstacked: Boolean
         get() = count > itemType.maxCount
 
+    inline val overstackedAndNotManageable: Boolean
+        get() = count > itemType.maxCount && count > 64
+
+
     final override fun toString() = "${count}x $itemType"
 
     final override fun equals(other: Any?): Boolean {